-arrange the projects folders -merge dbconnector with and without db name
