+++ src/main/java/net/education/kyivstar/Application.java
@@ -1,10 +1,9 @@
 package net.education.kyivstar;
 
 
+import net.education.kyivstar.services.db.DbConnector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.sql.SQLException;
 
 import static net.education.kyivstar.config.EducationEmbeddedMariaDb.startEmbeddedMariaDB;
 import static net.education.kyivstar.config.EducationEmbeddedMariaDb.stopEmbeddedMariaDB;
@@ -12,18 +11,9 @@ import static net.education.kyivstar.config.EducationEmbeddedMariaDb.stopEmbedde
 public class Application {
     private static final Logger logger = LoggerFactory.getLogger(Application.class);
 
-    public static void main(String[] args) throws SQLException {
-      // stopEmbeddedMariaDB();
-     /*   startEmbeddedMariaDB();
-        CreateSchema createSchema = new CreateSchema();
-        createSchema.createDataBase();
-        createSchema.createTables();
-        //stopEmbeddedMariaDB();*/
-      // final var f =  PasswordEncryptor.encrypt("password");
+    public static void main(String[] args) {
         startEmbeddedMariaDB();
         DbConnector dbConnector = new DbConnector();
-        dbConnector.connectMariaDb();
         stopEmbeddedMariaDB();
     }
-
 }
