--- src/main/java/net/education/kyivstar/premises/HumanService.java
@@ -1,128 +1,0 @@
-package net.education.kyivstar.premises;
-
-public class HumanService<Human> {
- /*   private List<Human> students;
-    private List<Human> revisers;
-
-    public HumanService() {
-        this.students = new ArrayList<>();
-        this.revisers = new ArrayList<>();
-    }
-
-
-    public List<T> findBySurname(String surname) {
-        List<T> result = new ArrayList<>();
-        for (T student : students) {
-            if (student.getSurname().equals(surname)) {
-                result.add(student);
-
-            }
-            System.out.println("student"+student);
-            System.out.println("surname"+surname);
-            System.out.println(" ddfsf "+student.equals(surname));
-        }
-        return result;
-    }
-
-
-   /*
-
-    private static List<FirstYearStudent> firstYearStudentList;
-
-
-    public static List<FirstYearStudent> createFirstYearStudents(int numberOfStudents) {
-        Faker faker = new Faker();
-
-        //   FirstYearStudent student = new FirstYearStudent(name,surname,age);
-        firstYearStudentList = new ArrayList<>();
-        for (int count = 0; count < numberOfStudents; count++) {
-            String name = faker.name().firstName();
-            String surname = faker.name().lastName();
-            int age = faker.number().numberBetween(16, 30);
-            FirstYearStudent firstYearStudent = new FirstYearStudent(name, surname, age);
-            firstYearStudentList.add(firstYearStudent);
-        }
-        System.out.println(firstYearStudentList);
-        return firstYearStudentList;
-    }
-
-
-    public static void createSecondYearStudents(int numberOfStudents) {
-        Faker faker = new Faker();
-
-        //   FirstYearStudent student = new FirstYearStudent(name,surname,age);
-        List<SecondYearStudent> secondYearStudentList = new ArrayList<>();
-        for (int count = 0; count < numberOfStudents; count++) {
-            String name = faker.name().firstName();
-            String surname = faker.name().lastName();
-            int age = faker.number().numberBetween(16, 30);
-            SecondYearStudent firstYearStudent = new SecondYearStudent(name, surname, age);
-            secondYearStudentList.add(firstYearStudent);
-        }
-        System.out.println(secondYearStudentList);
-    }
-
-    public static void createThirdYearStudents(int numberOfStudents) {
-        Faker faker = new Faker();
-
-        List<ThirdYearStudent> thirdYearStudentList = new ArrayList<>();
-        for (int count = 0; count < numberOfStudents; count++) {
-            String name = faker.name().firstName();
-            String surname = faker.name().lastName();
-            int age = faker.number().numberBetween(16, 30);
-            ThirdYearStudent thirdYearStudent = new ThirdYearStudent(name, surname, age);
-            thirdYearStudentList.add(thirdYearStudent);
-        }
-        System.out.println(thirdYearStudentList);
-    }
-
-    public static void createFourthYearStudents(int numberOfStudents) {
-        Faker faker = new Faker();
-
-        List<FourthYearStudent> fourthYearStudentList = new ArrayList<>();
-        for (int count = 0; count < numberOfStudents; count++) {
-            String name = faker.name().firstName();
-            String surname = faker.name().lastName();
-            int age = faker.number().numberBetween(16, 30);
-            FourthYearStudent fourthYearStudent = new FourthYearStudent(name, surname, age);
-            fourthYearStudentList.add(fourthYearStudent);
-        }
-        System.out.println(fourthYearStudentList);
-    }
-
-    public static void createFifthYearStudents(int numberOfStudents) {
-        Faker faker = new Faker();
-
-        List<FifthYearStudent> fifthYearStudentsList = new ArrayList<>();
-        for (int count = 0; count < numberOfStudents; count++) {
-            String name = faker.name().firstName();
-            String surname = faker.name().lastName();
-            int age = faker.number().numberBetween(16, 30);
-            FifthYearStudent fifthYearStudent = new FifthYearStudent(name, surname, age);
-            fifthYearStudentsList.add(fifthYearStudent);
-        }
-        System.out.println(fifthYearStudentsList);
-    }
-
-    public static void createRevisers(int numberOfReviser) {
-        Faker faker = new Faker();
-
-        List<Reviser> listOfRevisers = new ArrayList<>();
-        for (int count = 0; count < numberOfReviser; count++) {
-            String name = faker.name().firstName();
-            String surname = faker.name().lastName();
-            int age = faker.number().numberBetween(25, 80);
-            Reviser firstYearStudent = new Reviser(name, surname, age);
-            listOfRevisers.add(firstYearStudent);
-        }
-        System.out.println(listOfRevisers);
-    }*/
-
-}
-
-
-
-
-
-
-
