+++ src/main/java/net/education/kyivstar/repositories/StudentRepository.java
@@ -1,5 +1,8 @@
-package net.education.kyivstar;
+package net.education.kyivstar.repositories;
 
+import net.education.kyivstar.services.db.DbConnector;
+import net.education.kyivstar.services.user.UserService;
+import net.education.kyivstar.services.util.Utils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -10,21 +13,23 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class ReviserRepository extends DbConnector {
+public class StudentRepository extends DbConnector {
     private static final Logger logger = LoggerFactory.getLogger(UserService.class);
-
     private Connection conn = null;
 
-    private void openConnection() {
-        conn = connectMariaDb();
+    public StudentRepository() {
     }
 
-    public void addReviser(String surname, String name, int age) throws SQLException {
+    private void openConnection() {
+        conn = connectMariaDb(true);
+    }
+
+    public void addStudent(String surname, String name, int age) throws SQLException {
         openConnection();
         PreparedStatement ps = null;
 
         try {
-            ps = conn.prepareStatement("INSERT INTO REVISERS (SURNAME,NAME,AGE,ENTRY_DATE) VALUES(?,?,?,?)");
+            ps = conn.prepareStatement("INSERT INTO STUDENTS (SURNAME,NAME,AGE,ENTRY_DATE) VALUES(?,?,?,?)");
             ps.setString(1, surname);
             ps.setString(2, name);
             ps.setInt(3, age);
@@ -41,36 +46,13 @@ public class ReviserRepository extends DbConnector {
 
     }
 
-    public List<Object> extractAllRevisers() throws SQLException {
+    public List<Object> extractStudentsBySurname(String surname) throws SQLException {
 
         openConnection();
         PreparedStatement ps = null;
         List<Object> resultTasks = new ArrayList<>();
         try {
-            ps = conn.prepareStatement("SELECT SURNAME,NAME,AGE FROM REVISERS r");
-
-            ResultSet rs = ps.executeQuery();
-            while (rs.next()) {
-                resultTasks.add(rs.getString("SURNAME"));
-                resultTasks.add(rs.getString("NAME"));
-                resultTasks.add(rs.getInt("AGE"));
-            }
-        } catch (SQLException e) {
-            e.printStackTrace();
-        } finally {
-            ps.close();
-            conn.close();
-        }
-        return resultTasks;
-    }
-
-    public List<Object> extractRevisersBySurname(String surname) throws SQLException {
-
-        openConnection();
-        PreparedStatement ps = null;
-        List<Object> resultTasks = new ArrayList<>();
-        try {
-            ps = conn.prepareStatement("SELECT SURNAME,NAME,AGE FROM REVISERS r WHERE SURNAME =?");
+            ps = conn.prepareStatement("SELECT SURNAME,NAME,AGE FROM STUDENTS r WHERE SURNAME =?");
             ps.setString(1, surname);
             ps.executeUpdate();
 
@@ -80,6 +62,32 @@ public class ReviserRepository extends DbConnector {
                 resultTasks.add(rs.getString("NAME"));
                 resultTasks.add(rs.getInt("AGE"));
             }
+            System.out.println(ps);
+        } catch (SQLException e) {
+            e.printStackTrace();
+        } finally {
+            ps.close();
+            conn.close();
+        }
+        System.out.println(resultTasks);
+        return resultTasks;
+    }
+
+    public List<Object> extractAllStudents() throws SQLException {
+
+        openConnection();
+        PreparedStatement ps = null;
+        List<Object> resultTasks = new ArrayList<>();
+        try {
+            ps = conn.prepareStatement("" +
+                    "SELECT SURNAME,NAME,AGE, GROUP_ID FROM STUDENTS r");
+
+            ResultSet rs = ps.executeQuery();
+            while (rs.next()) {
+                resultTasks.add(rs.getString("SURNAME"));
+                resultTasks.add(rs.getString("NAME"));
+                resultTasks.add(rs.getInt("AGE"));
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         } finally {
@@ -89,11 +97,11 @@ public class ReviserRepository extends DbConnector {
         return resultTasks;
     }
 
-    public void updateReviserBySurname(String surname, String surnameToReplace, String nameToReplace, int ageToReplace) throws SQLException {
+    public void updateStudentBySurname(String surname, String surnameToReplace, String nameToReplace, int ageToReplace) throws SQLException {
         openConnection();
         PreparedStatement ps = null;
         try {
-            ps = conn.prepareStatement("UPDATE REVISERS SET SURNAME = ?, NAME =?, AGE =? WHERE SURNAME =?");
+            ps = conn.prepareStatement("UPDATE STUDENTS SET SURNAME = ?, NAME =?, AGE =? WHERE SURNAME =?");
             ps.setString(1, surnameToReplace);
             ps.setString(2, nameToReplace);
             ps.setInt(3, ageToReplace);
@@ -108,11 +116,11 @@ public class ReviserRepository extends DbConnector {
         }
     }
 
-    public void removeReviserBySurname(String surname) throws SQLException {
+    public void removeStudentBySurname(String surname) throws SQLException {
         openConnection();
         PreparedStatement ps = null;
         try {
-            ps = conn.prepareStatement("DELETE FROM REVISERS WHERE SURNAME =?");
+            ps = conn.prepareStatement("DELETE FROM STUDENTS WHERE SURNAME =?");
             ps.setString(1, surname);
             ps.executeUpdate();
         } catch (SQLException e) {
@@ -123,3 +131,4 @@ public class ReviserRepository extends DbConnector {
         }
     }
 }
+
