+++ src/main/java/net/education/kyivstar/services/user/ValidationService.java
@@ -1,4 +1,4 @@
-package net.education.kyivstar;
+package net.education.kyivstar.services.user;
 
 public class ValidationService {
