+++ src/main/resources/application.yml
@@ -1,7 +1,7 @@
 database:
   url: jdbc:mariadb://localhost
-  port: 3321
+  port: 3323
   username: 'user'
-  password: 4ZOhlGf6p2R+DLlghvYBgHPZk6Dbu5os
+  password: ENC(4ZOhlGf6p2R+DLlghvYBgHPZk6Dbu5os)
   dbName: 'EDUCATION'
   directory: "D:\\Education\\MariaBBB\\d"
