+++ src/main/resources/mariaDb/init.sql
@@ -1,9 +1,4 @@
 CREATE DATABASE IF NOT EXISTS EDUCATION;
 
-CREATE USER 'user'@'localhost' IDENTIFIED BY 'password';
-GRANT ALL PRIVILEGES ON EDUCATION.* TO 'user'@'localhost';
---#CREATE USER 'EDUCATION'@'%' IDENTIFIED BY 'EDUCATION';
--- # GRANT ALL PRIVILEGES ON *.* TO 'EDUCATION'@'%';
--- # FLUSH PRIVILEGES;
